Move to private implementation instead of helper functions
